requirements.txt
streamlit>=1.32.0

# MCP
mcp>=1.0.0

# LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0

# Tools
duckduckgo-search>=6.0.0
wikipedia>=1.4.0
requests>=2.31.0

# Pydantic
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0